\name{patPRO-package}
\alias{patPRO-package}
\alias{patPRO}
\docType{package}
\title{
Longitudinal Profiler of Microbiome Data
}
\description{
patPRO (short for Patient Profiler) is a package for formatting Qiime microbiome analysis output files for downstream analyses and generating powerful graphics for visualizing longitudinal datasets per patient or averages of patient cohorts.
}
\details{
\tabular{ll}{
Package: \tab patPRO\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2015-09-18\cr
License: \tab GPLv3\cr
}
An overview of how to use the package, including the most important functions, is included in the supplemental R notebook patPROExampleWorkflow.html.
}
\author{
Geoffrey Hannigan <ghanni@upenn.edu>, Loesche MA, Hodkinson BP, Mehta S, Elizabeth Grice <egrice@upenn.edu>
}
\references{
patPRO: An R package for the visualization of longitudinal microbiome data. [Submitted] Hannigan GD, Loesche MA, Hodkinson BP, Mehta S, Grice EA. Bioinformatics.
}
\keyword{ package }
