\name{PatProAlphaDiv}
\alias{PatProAlphaDiv}
\docType{data}
\title{
Test Alpha Diversity Data Used as an Example for patPRO.
}
\description{
This is the test alpha diversity data used as an example for patPRO.
The data is from Hannigan et al. 2014. "Culture-independent pilot 
study of microbiota colonizing open fractures and association with 
severity, mechanism, location, and complication from presentation to 
early outpatient follow-up". DOI: 10.1002/jor.22578
}

\usage{
data(PatProAlphaDiv)
}
\format{data frame}
\source{DOI: 10.1002/jor.22578}
\keyword{datasets}


\details{
\tabular{ll}{
Package: \tab patPRO\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2015-09-18\cr
License: \tab GPLv3\cr
}
An overview of how to use the package, including the most important functions, is included in the supplemental R notebook patPROExampleWorkflow.html.
}
\author{
Geoffrey Hannigan <ghanni@upenn.edu>, Loesche MA, Hodkinson BP, Mehta S, Elizabeth Grice <egrice@upenn.edu>
}
\references{
patPRO: An R package for the visualization of longitudinal microbiome data. Hannigan GD, Loesche MA, Hodkinson BP, Mehta S, Grice EA.
}
\keyword{ package }
